+++ cynic-introspection/tests/detection-tests.rs
@@ -1,12 +1,12 @@
 use cynic::GraphQlResponse;
-use cynic_introspection::{CapabilityDetectionQuery, SpecificationVersion};
+use cynic_introspection::{CapabilitiesQuery, SpecificationVersion};
 
 #[test]
 fn shapshot_test_detection_query() {
     use cynic::QueryBuilder;
 
-    insta::assert_snapshot!(CapabilityDetectionQuery::build(()).query, @r###"
-    query CapabilityDetectionQuery {
+    insta::assert_snapshot!(CapabilitiesQuery::build(()).query, @r###"
+    query CapabilitiesQuery {
       __type(name: "__Type") {
         fields(includeDeprecated: true) {
           name
@@ -19,9 +19,9 @@ fn shapshot_test_detection_query() {
 
 #[test]
 fn detection_query_with_2018_response() {
-    let capabilities = serde_json::from_str::<GraphQlResponse<CapabilityDetectionQuery>>(
-        include_str!("detection-responses/2018.json"),
-    )
+    let capabilities = serde_json::from_str::<GraphQlResponse<CapabilitiesQuery>>(include_str!(
+        "detection-responses/2018.json"
+    ))
     .unwrap();
 
     assert_eq!(
@@ -32,7 +32,7 @@ fn detection_query_with_2018_response() {
 
 #[test]
 fn detection_query_with_2021_response() {
-    let response = serde_json::from_str::<GraphQlResponse<CapabilityDetectionQuery>>(include_str!(
+    let response = serde_json::from_str::<GraphQlResponse<CapabilitiesQuery>>(include_str!(
         "detection-responses/2021.json"
     ))
     .unwrap();